// src/app/interfaces/metricas-resumen-incidencia.ts
export interface MetricasResumenIncidencia {
  totalResumen: number;
  totalValorizadoResumen: number;
  nuevasResumen: number;
  nuevasValorizadoResumen: number;
  enRevisionResumen: number;
  enRevisionValorizadoResumen: number;
  aprobadasResumen: number;
  aprobadasValorizadoResumen: number;
  rechazadasResumen: number;
  rechazadasValorizadoResumen: number;
  porcentajeNuevasResumen: number;
  porcentajeEnRevisionResumen: number;
  porcentajeAprobadasResumen: number;
  porcentajeRechazadasResumen: number;
}