// src/app/interfaces/metricas-bodega-incidencia.ts
export interface MetricasBodegaIncidencia {
    totalBodega: number;
    totalValorizadoBodega: number;
    nuevasBodega: number;
    nuevasValorizadoBodega: number;
    enRevisionBodega: number;
    enRevisionValorizadoBodega: number;
    aprobadasBodega: number;
    aprobadasValorizadoBodega: number;
    rechazadasBodega: number;
    rechazadasValorizadoBodega: number;
    porcentajeNuevasBodega: number;
    porcentajeEnRevisionBodega: number;
    porcentajeAprobadasBodega: number;
    porcentajeRechazadasBodega: number;
}